+++ CHANGELOG.md
@@ -6,6 +6,30 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html), starting with v1.1.0.
 
 
+## 1.2.1 - TBD
+
+### Added
+
+- Nothing.
+
+### Changed
+
+- Uses Trait for Scaffold
+- Added new layout for scaffolded views, named `scaffold` for easier separation
+
+### Deprecated
+
+- Nothing.
+
+### Removed
+
+- Nothing.
+
+### Fixed
+
+- Nothing.
+
+
 ## 1.2.0 - 2018-04-30
 
 ### Added